WIP: Move Editor connect to kapture-backend
