+++ ui/src/components/index.js
@@ -3,3 +3,4 @@ export { ActivitySearch } from './activity/ActivitySearch';
 export { ActivityHeader } from './activity/ActivityHeader';
 export { PlateSizeForm } from './activity/PlateSizeForm';
 export { Print } from './print';
+export { Editor } from './editor/Editor';
