+++ ui/src/implementations/kapture-backend/components/index.js
@@ -4,3 +4,4 @@ export { ActivityHeader } from './ActivityHeaderConnected';
 export { ActivityDetails } from './ActivityDetails';
 export { PlateSizeForm } from '../../../components';
 export { Print } from './PrintConnected';
+export { Editor } from './EditorConnected';
