+++ ui/src/implementations/kapture-backend/routes/activities/Activities.js
@@ -5,8 +5,8 @@ import { Route, Switch, matchPath } from 'react-router-dom';
 
 import { ActivityHeader } from '../../components';
 import { ActivityDetails } from '../../components';
-// import { Editor } from '../../../../components/editor/Editor';
-import { EditorActions } from '../../../../components/editor/EditorActions';
+import { Editor } from '../../components';
+// import { EditorActions } from '../../../../components/editor/EditorActions';
 import { Print } from '../../components';
 import { PrintActions } from '../../../../components/print/PrintActions';
 import { CompletedModal } from './CompletedModal';
@@ -123,7 +123,8 @@ export class Activities extends Component {
     } else if (initialized) {
       if (this.matchEditorPath() && editorInitialized) {
         actions = (
-          <EditorActions onMarkAsCompleted={this.handleMarkAsCompleted} />
+          <div />
+          // <EditorActions onMarkAsCompleted={this.handleMarkAsCompleted} />
         );
       } else if (this.matchPrintPath() && printInitialized) {
         actions = <PrintActions contentRef={this.state.contentRef} />;
@@ -141,7 +142,7 @@ export class Activities extends Component {
                 );
               }}
             />
-            {/* <Route path={`${match.path}/editor`} component={Editor} /> */}
+            <Route path={`${match.path}/editor`} component={Editor} />
             <Route
               path={`${match.path}/print`}
               render={routeProps => (
